fix: remove dead cluster panel
Closes #15
